+++ src/grib1/implementation/LambertConformalImpl.h
@@ -28,9 +28,9 @@ class LambertConformalImpl : public LambertConformal
     bool                reverseXDirection() const;
     bool                reverseYDirection() const;
 
-    bool                getProperty(uint propertyId,long long& value);
+    bool                getProperty(uint propertyId,Int64& value);
     void                getProperties(T::PropertySettingVec& properties);
-    bool                setProperty(uint propertyId,long long value);
+    bool                setProperty(uint propertyId,Int64 value);
 
     void                initSpatialReference();
     void                read(MemoryReader& memoryReader);
