+++ src/grib1/implementation/MercatorImpl.cpp
@@ -406,14 +406,14 @@ void MercatorImpl::getProperties(T::PropertySettingVec& properties)
 
 
 
-/*! \brief The method is used for fetching a (long long ) value for the property according to the property id.
+/*! \brief The method is used for fetching a (Int64 ) value for the property according to the property id.
 
         \param propertyId  The (numeric) identifier of the requested property.
         \param value       The value of the requested property is returned in this parameter.
         \return            The method returns true if the value of the requested property was found.
 */
 
-bool MercatorImpl::getProperty(uint propertyId,long long& value)
+bool MercatorImpl::getProperty(uint propertyId,Int64& value)
 {
   try
   {
@@ -452,14 +452,14 @@ bool MercatorImpl::getProperty(uint propertyId,long long& value)
 
 
 
-/*! \brief The method is used for setting a (long long) value for the property according to the property id.
+/*! \brief The method is used for setting a (Int64) value for the property according to the property id.
 
         \param propertyId  The (numeric) identifier of the requested property.
         \param value       The value of the property to be set.
         \return            The method returns true if the value of the requested property was set.
 */
 
-bool MercatorImpl::setProperty(uint propertyId,long long value)
+bool MercatorImpl::setProperty(uint propertyId,Int64 value)
 {
   try
   {
