+++ src/grib2/IndicatorSection.cpp
@@ -165,14 +165,14 @@ void IndicatorSection::setMessagePtr(Message *message)
 
 
 
-/*! \brief The method is used for fetching a (long long ) value for the property according to the property id.
+/*! \brief The method is used for fetching a (Int64 ) value for the property according to the property id.
 
         \param propertyId  The (numeric) identifier of the requested property.
         \param value       The value of the requested property is returned in this parameter.
         \return            The method returns true if the value of the requested property was found.
 */
 
-bool IndicatorSection::getProperty(uint propertyId,long long& value)
+bool IndicatorSection::getProperty(uint propertyId,Int64& value)
 {
   try
   {
@@ -219,14 +219,14 @@ void IndicatorSection::getProperties(T::PropertySettingVec& properties)
 
 
 
-/*! \brief The method is used for setting a (long long) value for the property according to the property id.
+/*! \brief The method is used for setting a (Int64) value for the property according to the property id.
 
         \param propertyId  The (numeric) identifier of the requested property.
         \param value       The value of the property to be set.
         \return            The method returns true if the value of the requested property was set.
 */
 
-bool IndicatorSection::setProperty(uint propertyId,long long value)
+bool IndicatorSection::setProperty(uint propertyId,Int64 value)
 {
   try
   {
