+++ src/grib2/LocalSection.h
@@ -61,9 +61,9 @@ class LocalSection : public GRID::MessageSection
     void            setMessagePtr(Message *message);
     void            setData(T::Data_ptr data,std::size_t size);
 
-    bool            getProperty(uint propertyId,long long& value);
+    bool            getProperty(uint propertyId,Int64& value);
     void            getProperties(T::PropertySettingVec& properties);
-    bool            setProperty(uint propertyId,long long value);
+    bool            setProperty(uint propertyId,Int64 value);
 
     void            read(MemoryReader& memoryReader);
     void            write(DataWriter& dataWriter);
